+++ astromodels/functions/template_model.py
@@ -1,4 +1,5 @@
-import collections
+import logging
+
 import gc
 import os
 import re
@@ -17,9 +18,8 @@ from astromodels.core.parameter import Parameter
 from astromodels.functions.function import Function1D, FunctionMeta
 from astromodels.utils import get_user_data_path
 from astromodels.utils.file_utils import copy_if_needed
-from astromodels.utils.logging import setup_logger
 
-log = setup_logger(__name__)
+log = logging.getLogger(__name__)
 
 # A very small number which will be substituted to zero during the construction
 # of the templates
@@ -136,9 +136,7 @@ class TemplateModelFactory(object):
 
         # We create a dictionary which will contain the grid for each parameter
 
-        self._parameters_grids: Dict[str, Optional[np.ndarray]] = (
-            collections.OrderedDict()
-        )
+        self._parameters_grids: Dict[str, Optional[np.ndarray]] = dict()
 
         for parameter_name in names_of_parameters:
             self._parameters_grids[parameter_name] = None
@@ -457,7 +455,7 @@ class TemplateFile:
 
             grid = f["grid"][()]
 
-            parameters = collections.OrderedDict()
+            parameters = dict()
 
             for k in parameter_order:
                 parameters[k] = f["parameters"][k][()]
@@ -547,7 +545,7 @@ class TemplateModel(Function1D, metaclass=FunctionMeta):
         except Exception:
             raise InvalidTemplateModelFile()
 
-        self._parameters_grids = collections.OrderedDict()
+        self._parameters_grids = dict()
 
         for key in template_file.parameter_order:
             try:
@@ -579,7 +577,7 @@ class TemplateModel(Function1D, metaclass=FunctionMeta):
 
         # Make the dictionary of parameters
 
-        function_definition = collections.OrderedDict()
+        function_definition = dict()
 
         function_definition["description"] = description
 
@@ -587,7 +585,7 @@ class TemplateModel(Function1D, metaclass=FunctionMeta):
 
         # Now build the parameters according to the content of the parameter grid
 
-        parameters = collections.OrderedDict()
+        parameters = dict()
 
         parameters["K"] = Parameter("K", 1.0)
         parameters["scale"] = Parameter("scale", 1.0)
@@ -963,7 +961,7 @@ def convert_old_table_model(model_name: str):
     with HDFStore(filename_sanitized) as store:
         data_frame = store["data_frame"]
 
-        parameters_grids = collections.OrderedDict()
+        parameters_grids = dict()
 
         processed_parameters = 0
 
@@ -1015,7 +1013,7 @@ def convert_old_table_model(model_name: str):
 
         # Make the dictionary of parameters
 
-        function_definition = collections.OrderedDict()
+        function_definition = dict()
 
         function_definition["description"] = description