+++ astromodels/sources/extended_source.py
@@ -1,4 +1,4 @@
-import collections
+import logging
 
 import astropy.units as u
 import numpy as np
@@ -8,10 +8,10 @@ from astromodels.core.tree import Node
 from astromodels.core.units import get_units
 from astromodels.functions import Constant
 from astromodels.sources.source import Source, SourceType
-from astromodels.utils.logging import setup_logger
+
 from astromodels.utils.pretty_list import dict_to_list
 
-log = setup_logger(__name__)
+log = logging.getLogger(__name__)
 
 
 class ExtendedSource(Source, Node):
@@ -294,7 +294,7 @@ class ExtendedSource(Source, Node):
 
         :return:
         """
-        free_parameters = collections.OrderedDict()
+        free_parameters = dict()
 
         for component in list(self._components.values()):
 
@@ -320,7 +320,7 @@ class ExtendedSource(Source, Node):
 
         :return:
         """
-        all_parameters = collections.OrderedDict()
+        all_parameters = dict()
 
         for component in list(self._components.values()):
 
@@ -343,13 +343,13 @@ class ExtendedSource(Source, Node):
 
         # Make a dictionary which will then be transformed in a list
 
-        repr_dict = collections.OrderedDict()
+        repr_dict = dict()
 
         key = "%s (extended source)" % self.name
 
-        repr_dict[key] = collections.OrderedDict()
+        repr_dict[key] = dict()
         repr_dict[key]["shape"] = self._spatial_shape.to_dict(minimal=True)
-        repr_dict[key]["spectrum"] = collections.OrderedDict()
+        repr_dict[key]["spectrum"] = dict()
 
         for component_name, component in list(self.components.items()):
             repr_dict[key]["spectrum"][component_name] = component.to_dict(minimal=True)