+++ astromodels/utils/logging.py
@@ -51,8 +51,8 @@ def get_path_of_log_dir():
 
     # Create it if doesn't exist
     if not user_log.exists():
-
-        user_log.mkdir(parents=True)
+        # still need to catch this for parallelized tests
+        user_log.mkdir(parents=True, exist_ok=True)
 
     return user_log
 
@@ -202,7 +202,23 @@ def setup_logger(name):
 
     log.addHandler(astromodels_usr_log_handler)
 
-    # we do not want to duplicate teh messages in the parents
-    log.propagate = False
-
     return log
+
+
+# Capture all startup warnings and log them on demand
+
+
+_startup_warnings = []
+
+
+def add_startup_warning(logger, msg, level=logging.WARNING):
+    fn, lno, func, sinfo = logger.findCaller(stacklevel=2)
+    record = logger.makeRecord(
+        logger.name, level, fn, lno, msg, (), None, func=func, sinfo=sinfo
+    )
+    _startup_warnings.append(record)
+
+
+def log_astromodels_startup_warnings(logger):
+    for w in _startup_warnings:
+        logger.handle(w)